moves card-header component to nested folder and breaks into separate files
