+++ libs/lorenzo/ui/src/lib/cards/card/card-header/card-header.component.html
@@ -1,0 +1,8 @@
+<div>
+  <ng-content></ng-content>
+</div>
+<fa-icon *ngIf="canFavorite"
+  class="favorite"
+  [icon]="favorite ? favoriteIcon : notFavoriteIcon"
+  (click)="favoriteChange.emit(!favorite)"
+></fa-icon>
